[finance_report_infra] fix(frontend): move analytics-env helper out of layout.tsx (Next layout export rule)
Next.js rejects arbitrary exports from a Layout file ("does not match the
required types of a Next.js Layout"), which broke `next build`. Move
analyticsClientIdMissingInDeployedEnv into src/lib/analytics-env.ts; layout.tsx
and the test import it from there. Behavior unchanged.

Verified: next build compiles, vitest 3 passed, tsc clean.
